+++ R/rescale.R
@@ -1,4 +1,4 @@
-# NORMALIZING
+# SCALING
 #' @include AllGenerics.R
 NULL
 
@@ -11,6 +11,7 @@ setMethod(
   signature = signature(x = "numeric", y = "numeric"),
   definition = function(x, y, method = c("rectangle", "trapezoid"), ...) {
     ## Validation
+    assert_length(y, length(x))
     method <- match.arg(method, several.ok = FALSE)
 
     ## Get method
@@ -49,7 +50,9 @@ setMethod(
   f = "rescale_total",
   signature = signature(x = "numeric", y = "numeric"),
   definition = function(x, y, total = 1) {
-    y <- (y * total) / sum(y)
+    assert_length(y, length(x))
+
+    y <- (y * total) / sum(abs(y))
     xy <- list(x = x, y = y)
     xy
   }
@@ -76,6 +79,7 @@ setMethod(
   signature = signature(x = "numeric", y = "numeric"),
   definition = function(x, y, min = 0, max = 1) {
     ## Validation
+    assert_length(y, length(x))
     if (min > max) {
       msg <- sprintf("%s (%g) must be lower than %s (%g).",
                      sQuote("min"), min, sQuote("max"), max)
@@ -154,6 +158,8 @@ setMethod(
   f = "rescale_transform",
   signature = signature(x = "numeric", y = "numeric"),
   definition = function(x, y, f, ...) {
+    assert_length(y, length(x))
+
     y <- f(y, ...)
     xy <- list(x = x, y = y)
     xy
@@ -169,5 +175,35 @@ setMethod(
   definition = function(x, f, ...) {
     xy <- grDevices::xy.coords(x)
     methods::callGeneric(x = xy$x, y = xy$y, f = f, ...)
+  }
+)
+
+# SNV ==========================================================================
+#' @export
+#' @rdname rescale_snv
+#' @aliases rescale_snv,numeric,numeric-method
+setMethod(
+  f = "rescale_snv",
+  signature = signature(x = "numeric", y = "numeric"),
+  definition = function(x, y, ...) {
+    ## Validation
+    assert_length(y, length(x))
+
+    y <- (y - mean(y)) / stats::sd(y)
+
+    xy <- list(x = x, y = y)
+    xy
+  }
+)
+
+#' @export
+#' @rdname rescale_snv
+#' @aliases rescale_snv,ANY,missing-method
+setMethod(
+  f = "rescale_snv",
+  signature = signature(x = "ANY", y = "missing"),
+  definition = function(x, ...) {
+    xy <- grDevices::xy.coords(x)
+    methods::callGeneric(x = xy$x, y = xy$y, ...)
   }
 )